services/orders/requirements.txt
Flask
Flask-CORS
Flask-Swagger-UI
jsonschema
gunicorn
python-dotenv